Update Risk working

+++ backend/infosafe_backend/src/main/java/com/fragile/infosafe/primary/model/Risk.java
@@ -14,7 +14,7 @@ import lombok.NoArgsConstructor;
 @Table(name="risks")
 public class Risk {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long risk_id;
+    private int risk_id;
     private String risk_name;
     private String impact_rating;
     private String probability_rating;
